+++ builder/definitions/index.ts
@@ -4,6 +4,7 @@ export interface IBuilderContext {
   stylesheets: string[];
   scripts: string[];
   html: string[];
+  other: string[];
   options: IBuilderOptions;
 }