+++ CHANGELOG.md
@@ -5,6 +5,10 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 
+## WIP
+### Added
+- `[compilers]` Added copy compiler for if no compiler is found for files.
+
 ## 0.5.6 - 6. August 2020
 ### Added
 - `[config]` Option for not clearing dist dir when building